Add FakePyDevd.wait_for_command().
